requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
line-bot-sdk==2.0.1
google-cloud-firestore==2.3.4
google-cloud-language==2.3.0
feedparser==6.0.8
requests==2.26.0
google-generativeai==0.3.1